+++ app/connectors_service/tests/sources/test_mssql.py
@@ -13,7 +13,7 @@ from connectors_sdk.filtering.validation import Filter, SyncRuleValidationResult
 from sqlalchemy.engine import Engine
 from sqlalchemy.exc import ProgrammingError
 
-from connectors.sources.mssql import (
+from app.connectors_service.connectors.sources.mssql.mssql import (
     MSSQLAdvancedRulesValidator,
     MSSQLDataSource,
     MSSQLQueries,
